Refactor command execution to inherit parent fds
Instead of executing the command and then printing the output, specify
that the child process should inhert the parent's stdout/err file
descriptors for more natural output display. Also change from using a
`match` to using an `if let` because now we don't need to take action on
the successful case.